pyproject.toml
[tool.poetry]
name = "invicofapy"
version = "0.1.0"
description = ""
authors = ["Fernando Corrales <fscpython@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
pytest-playwright = "^0.7.0"
numpy = "^2.2.3"
pandas = "^2.2.3"
xlrd = "^2.0.1"
motor = "^3.7.0"
pydantic-settings = "^2.9.1"
fastapi-jwt = {extras = ["authlib"], version = "^0.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-mongo = "^3.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
ptpython = "^3.0.29"
openpyxl = "^3.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "src"